fix: ajustando lint
